requirements.txt
pandas==1.4.0
requests==2.26.0
PyYAML~=6.0